requirements.txt
pyautogen>=0.2.34
python-dotenv>=1.0.0
